proof: restore endpoint-referenced SEA3 covariance ceiling
